+++ packages/core/test/bundling/forms/bundle.golden_symbols.json
@@ -1629,9 +1629,6 @@
     "name": "u"
   },
   {
-    "name": "unimplemented"
-  },
-  {
     "name": "unwrapRNode"
   },
   {
