+++ packages/forms/src/directives/ng_control.ts
@@ -6,15 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-
 import {AbstractControlDirective} from './abstract_control_directive';
 import {ControlContainer} from './control_container';
 import {ControlValueAccessor} from './control_value_accessor';
-import {AsyncValidator, AsyncValidatorFn, Validator, ValidatorFn} from './validators';
 
-function unimplemented(): any {
-  throw new Error('unimplemented');
-}
 
 /**
  * @description
@@ -43,42 +38,6 @@ export abstract class NgControl extends AbstractControlDirective {
    * The value accessor for the control
    */
   valueAccessor: ControlValueAccessor|null = null;
-
-  /**
-   * @description
-   * The uncomposed array of synchronous validators for the control
-   *
-   * @internal
-   */
-  _rawValidators: Array<Validator|ValidatorFn> = [];
-
-  /**
-   * @description
-   * The uncomposed array of async validators for the control
-   *
-   * @internal
-   */
-  _rawAsyncValidators: Array<AsyncValidator|AsyncValidatorFn> = [];
-
-  /**
-   * @description
-   * The registered synchronous validator function for the control
-   *
-   * @throws An exception that this method is not implemented
-   */
-  get validator(): ValidatorFn|null {
-    return <ValidatorFn>unimplemented();
-  }
-
-  /**
-   * @description
-   * The registered async validator function for the control
-   *
-   * @throws An exception that this method is not implemented
-   */
-  get asyncValidator(): AsyncValidatorFn|null {
-    return <AsyncValidatorFn>unimplemented();
-  }
 
   /**
    * @description
