+++ packages/forms/src/directives/ng_model_group.ts
@@ -62,8 +62,8 @@ export class NgModelGroup extends AbstractFormGroupDirective implements OnInit, 
       @Optional() @Self() @Inject(NG_ASYNC_VALIDATORS) asyncValidators: any[]) {
     super();
     this._parent = parent;
-    this._validators = validators;
-    this._asyncValidators = asyncValidators;
+    this._setValidators(validators);
+    this._setAsyncValidators(asyncValidators);
   }
 
   /** @internal */
